Split actions component and fix scroll.
